Added temp photos to ExamCamera
Also changed the gridView and ViewModel

+++ app/src/main/java/br/com/graest/retinografo/ui/screens/exam/ExamDataViewModel.kt
@@ -25,20 +25,28 @@ class ExamDataViewModel(
     private val examDataDao: ExamDataDao
 ) : ViewModel() {
 
-    private val _bitmaps = MutableStateFlow<List<Bitmap>>(emptyList())
-    val bitmaps = _bitmaps.asStateFlow()
+//    private val _bitmaps = MutableStateFlow<List<Bitmap>>(emptyList())
+//    val bitmaps = _bitmaps.asStateFlow()
+//
+//
+//    init {
+//        viewModelScope.launch {
+//            examDataDao.getExamData()
+//                .collect { entities ->
+//                    val loadedBitmaps = entities.map { entity ->
+//                        byteArrayToBitmap(entity.image)
+//                    }
+//                    _bitmaps.value = loadedBitmaps
+//                }
+//        }
+//    }
 
-    init {
-        viewModelScope.launch {
-            examDataDao.getExamData()
-                .collect { entities ->
-                    val loadedBitmaps = entities.map { entity ->
-                        byteArrayToBitmap(entity.image)
-                    }
-                    _bitmaps.value = loadedBitmaps
-                }
-        }
-    }
+
+    private val _capturedImagePaths = MutableStateFlow<List<String>>(emptyList())
+    val capturedImagePaths: StateFlow<List<String>> = _capturedImagePaths.asStateFlow()
+
+    private val _errorMessage = MutableStateFlow<String?>(null)
+    val errorMessage: StateFlow<String?> = _errorMessage.asStateFlow()
 
     private val _examsData = examDataDao.getExamData()
 
@@ -56,11 +64,19 @@ class ExamDataViewModel(
     fun onEvent(event: ExamDataEvent) {
         when (event) {
             ExamDataEvent.DeleteExamData -> {
-                val examData = ExamData(
-                    image = examDataState.value.image
-                )
+                val examData = examDataState.value.patientData?.let {
+                    ExamData(
+                        image1 = examDataState.value.image1,
+                        image2 = examDataState.value.image2,
+                        image3 = examDataState.value.image3,
+                        image4 = examDataState.value.image4,
+                        patientId = it.patientId
+                    )
+                }
                 viewModelScope.launch {
-                    examDataDao.deleteExamData(examData)
+                    if (examData != null) {
+                        examDataDao.deleteExamData(examData)
+                    }
                 }
             }
             is ExamDataEvent.DeleteExamDataById -> {
@@ -123,13 +139,21 @@ class ExamDataViewModel(
         }
     }
 
-    fun onTakePhoto(bitmap: Bitmap) {
-        viewModelScope.launch {
-            val image = ExamData(image = bitmapToByteArray(bitmap))
-            examDataDao.insertExam(image)
-            // Update the _bitmaps state flow
-            _bitmaps.value = _bitmaps.value + bitmap
-        }
+//    fun onTakePhoto(bitmap: Bitmap) {
+//        viewModelScope.launch {
+//            val image = ExamData(image = bitmapToByteArray(bitmap))
+//            examDataDao.insertExam(image)
+//            // Update the _bitmaps state flow
+//            _bitmaps.value = _bitmaps.value + bitmap
+//        }
+//    }
+
+    fun addImagePath(path: String) {
+        _capturedImagePaths.value = _capturedImagePaths.value + path
+    }
+
+    fun setErrorMessage(message: String?) {
+        _errorMessage.value = message
     }
